requirements.txt
Flask>=3.0.3,<3.1
Pillow>=11.0.0,<11.1